Refactors Supermarket section of Quality Metrics

cleans up code based on feedback
