+++ src/fieri/app/models/license_worker.rb
@@ -14,7 +14,7 @@ class LicenseWorker
   def perform(version_json, cookbook_name)
     version_info = JSON.parse(version_json)
     Net::HTTP.post_form(
-      URI.parse("#{ENV['FIERI_SUPERMARKET_ENDPOINT']}/api/v1/cookbook-versions/license_evaluation"),
+      URI.parse("#{ENV['FIERI_SUPERMARKET_ENDPOINT']}/api/v1/quality_metrics/license_evaluation"),
       fieri_key: ENV['FIERI_KEY'],
       cookbook_name: cookbook_name,
       cookbook_version: version_info['version'],
