+++ src/fieri/spec/models/collaborator_worker_spec.rb
@@ -4,7 +4,7 @@ describe CollaboratorWorker do
   let(:cw) { CollaboratorWorker.new }
 
   before do
-    stub_request(:post, "#{ENV['FIERI_SUPERMARKET_ENDPOINT']}/api/v1/cookbook-versions/collaborators_evaluation").
+    stub_request(:post, "#{ENV['FIERI_SUPERMARKET_ENDPOINT']}/api/v1/quality_metrics/collaborators_evaluation").
       to_return(status: 200, body: '', headers: {})
   end
 
@@ -14,7 +14,7 @@ describe CollaboratorWorker do
     it 'posts a passing metric' do
       cw.perform(cookbook_json_response)
 
-      assert_requested(:post, "#{ENV['FIERI_SUPERMARKET_ENDPOINT']}/api/v1/cookbook-versions/collaborators_evaluation", times: 1) do |req|
+      assert_requested(:post, "#{ENV['FIERI_SUPERMARKET_ENDPOINT']}/api/v1/quality_metrics/collaborators_evaluation", times: 1) do |req|
         expect(req.body).to include("cookbook_name=#{JSON.parse(cookbook_json_response)['name']}")
         expect(req.body).to include('collaborator_failure=false')
         expect(req.body).to include('collaborator_feedback=passed')
@@ -28,7 +28,7 @@ describe CollaboratorWorker do
     it 'posts a failing metric' do
       cw.perform(cookbook_json_response)
 
-      assert_requested(:post, "#{ENV['FIERI_SUPERMARKET_ENDPOINT']}/api/v1/cookbook-versions/collaborators_evaluation", times: 1) do |req|
+      assert_requested(:post, "#{ENV['FIERI_SUPERMARKET_ENDPOINT']}/api/v1/quality_metrics/collaborators_evaluation", times: 1) do |req|
         expect(req.body).to include("cookbook_name=#{JSON.parse(cookbook_json_response)['name']}")
         expect(req.body).to include('collaborator_failure=true')
         expect(req.body).to include('collaborator_feedback=Failure')
